Moved mock classes to a separate package since MockVaadin.kt was getting too big
