+++ karibu-testing-v10/kt10-testrun-vaadin18/src/test/kotlin/MockLitTemplateParser.kt
@@ -1,7 +1,7 @@
 package com.vaadin.flow.component.littemplate;
 
 import com.github.mvysny.kaributesting.v10.mock.MockInstantiator
-import com.github.mvysny.kaributesting.v10.MockNpmTemplateParser
+import com.github.mvysny.kaributesting.v10.mock.MockNpmTemplateParser
 import com.vaadin.flow.component.polymertemplate.TemplateParser
 import com.vaadin.flow.di.Instantiator
